backend/app/requirements.txt
elasticsearch==8.10.1
fastapi==0.106.0
pydantic==1.10.13
python-dotenv==1.0.0
SQLAlchemy==2.0.23
uvicorn==0.25.0
asyncpg==0.29.0